+++ Woof/Woof/NetworkLayer/Endpoint/WoofAppEndpoint+Path.swift
@@ -1,6 +1,8 @@
 extension WoofAppEndpoint {
     /// Represents endpoint paths for the application.
     enum Path {
+        /// The path to get sitters.
+        static let getAllSitters = "pet_sitters/all"
         /// The path for adding a new pet sitter.
         static let addNewSitter = "pet_sitters/new"
     }