+++ Woof/Woof/NetworkLayer/Endpoint/WoofAppEndpoint.swift
@@ -1,6 +1,8 @@
 import NetworkService
 ///  Defines the production endpoints for the Woof app's API.
 enum WoofAppEndpoint {
+    /// Receiving all sitters from the server.
+    case getAllSitters
     /// Adding a new pet sitter.
     case addNewSitter(BodyParameters?)
 }